implement limits for binomial test
these limits describe the upper and lower bounds of the test. Since we
work under the assumption that zero effect size p distributions will be
uniform, we can actually test a window of any size.

This commit includes adding the limits to the function, as well as a
function to check the limits, and lots of test to make sure both of
those work as expected.