split binonmial tests
we now have three tests, with associated test functions.

The sns test is from the p-curve original paper (two one-tailed tests).
The 'all' test is a version of this, but it's a single two-tailed test.
The 'bias' test is a binomial test specifically for publication bias
and/or p-hacking, comparing the upper two bins of the distribution
looking for an uptick.